# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /box:
    get:
      operationId: listBoxes
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      - description: Include verbose output
        explode: true
        in: query
        name: verbose
        required: false
        schema:
          type: boolean
        style: form
      - description: JSON encoded labels to filter by
        explode: true
        in: query
        name: labels
        required: false
        schema:
          example: "{\"label1\": \"value1\", \"label2\": \"value2\"}"
          type: string
        style: form
      - description: Include errored and deleted boxes
        explode: true
        in: query
        name: includeErroredDeleted
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Box"
                type: array
          description: List of all boxes
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List all boxes
      tags:
      - box
...
  /box/{boxIdOrName}:
    get:
      operationId: getBox
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-BoxLite-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      - description: ID or name of the box
        explode: false
        in: path
        name: boxIdOrName
        required: true
        schema:
          type: string
        style: simple
      - description: Include verbose output
        explode: true
        in: query
        name: verbose
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
          description: Box details
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get box details
      tags:
      - box
...
components:
  schemas:
...
    BoxState:
      description: The state of the box
      enum:
      - creating
      - restoring
      - destroyed
      - destroying
      - started
      - stopped
      - starting
      - stopping
      - error
      - unknown
      - archived
      - archiving
      - resizing
      type: string
    BoxDesiredState:
      description: The desired state of the box
      enum:
      - destroyed
      - started
      - stopped
      - resized
      type: string
    BoxVolume:
      example:
        volumeId: volume123
        mountPath: /data
        subpath: users/alice
      properties:
        volumeId:
          description: The ID of the volume
          example: volume123
          type: string
        mountPath:
          description: The mount path for the volume
          example: /data
          type: string
        subpath:
          description: "Optional subpath within the volume to mount. When specified,\
            \ only this S3 prefix will be accessible. When omitted, the entire volume\
            \ is mounted."
          example: users/alice
          type: string
      required:
      - mountPath
      - volumeId
      type: object
    Box:
      example:
        id: aB3cD4eF5gH6
        organizationId: organization123
        name: MyBox
        user: boxlite
        env:
          NODE_ENV: production
        labels:
          boxlite.io/public: "true"
        public: false
        networkBlockAll: false
        networkAllowList: "192.168.1.0/16,10.0.0.0/24"
        target: local
        image: boxlite/base
        cpu: 2
        gpu: 0
        memory: 4
        disk: 10
        state: creating
        desiredState: destroyed
        errorReason: The box is not running
        recoverable: true
        autoPauseInterval: 900
        autoDeleteInterval: 604800
        autoResumeEnabled: true
        volumes:
        - volumeId: volume123
          mountPath: /data
          subpath: users/alice
        - volumeId: volume123
          mountPath: /data
          subpath: users/alice
        createdAt: 2024-10-01T12:00:00Z
        updatedAt: 2024-10-01T12:00:00Z
        class: small
        daemonVersion: 1.0.0
        runnerId: runner123
        toolboxProxyUrl: https://proxy.app.boxlite.io/toolbox
      properties:
        id:
          description: The public 12-character Box ID
          example: aB3cD4eF5gH6
          type: string
        organizationId:
          description: The organization ID of the box
          example: organization123
          type: string
        name:
          description: The name of the box
          example: MyBox
          type: string
        user:
          description: The user associated with the project
          example: boxlite
          type: string
        env:
          additionalProperties:
            type: string
          description: Environment variables for the box
          example:
            NODE_ENV: production
          type: object
        labels:
          additionalProperties:
            type: string
          description: Labels for the box
          example:
            boxlite.io/public: "true"
          type: object
        public:
          description: Whether the box http preview is public
          example: false
          type: boolean
        networkBlockAll:
          description: Whether to block all network access for the box
          example: false
          type: boolean
        networkAllowList:
          description: Comma-separated list of allowed CIDR network addresses for
            the box
          example: "192.168.1.0/16,10.0.0.0/24"
          type: string
        target:
          description: The target environment for the box
          example: local
          type: string
        image:
          description: The image used for the box
          example: boxlite/base
          type: string
        cpu:
          description: The CPU quota for the box
          example: 2
          type: number
        gpu:
          description: The GPU quota for the box
          example: 0
          type: number
        memory:
          description: The memory quota for the box
          example: 4
          type: number
        disk:
          description: The disk quota for the box
          example: 10
          type: number
        state:
          allOf:
          - $ref: "#/components/schemas/BoxState"
          description: The state of the box
          example: creating
        desiredState:
          allOf:
          - $ref: "#/components/schemas/BoxDesiredState"
          description: The desired state of the box
          example: destroyed
        errorReason:
          description: The error reason of the box
          example: The box is not running
          type: string
        recoverable:
          description: Whether the box error is recoverable.
          example: true
          type: boolean
        autoPauseInterval:
          description: Auto-pause interval in seconds (0 means disabled)
          example: 900
          type: integer
        autoDeleteInterval:
          description: Auto-delete interval in seconds (0 means disabled)
          example: 604800
          type: integer
        autoResumeEnabled:
          description: Whether the box should be automatically resumed on proxy access
          example: true
          type: boolean
        volumes:
          description: Array of volumes attached to the box
          items:
            $ref: "#/components/schemas/BoxVolume"
          type: array
        createdAt:
          description: The creation timestamp of the box
          example: 2024-10-01T12:00:00Z
          type: string
        updatedAt:
          description: The last update timestamp of the box
          example: 2024-10-01T12:00:00Z
          type: string
        class:
          deprecated: true
          description: The class of the box
          enum:
          - small
          - medium
          - large
          example: small
          type: string
        daemonVersion:
          description: The version of the daemon running in the box
          example: 1.0.0
          type: string
        runnerId:
          description: The runner ID of the box
          example: runner123
          type: string
        toolboxProxyUrl:
          description: The toolbox proxy URL for the box
          example: https://proxy.app.boxlite.io/toolbox
          type: string
      required:
      - cpu
      - disk
      - env
      - gpu
      - id
      - labels
      - memory
      - name
      - networkBlockAll
      - organizationId
      - public
      - target
      - toolboxProxyUrl
      - user
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect